requirements.txt
python-dotenv>=1.0.0
httpx[http2]>=0.27.0
h2>=4.1.0
pydantic>=2.0.0
streamlit>=1.41.0
torch>=2.4.0
transformers>=4.46.0
